requirements.txt
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
matplotlib
pdfplumber
openpyxl>=3.1.0
lxml
python-dateutil>=2.8.0
yfinance>=0.2.0
firebase-admin>=6.2.0
